backend/requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9  # For PostgreSQL
# or
# mysqlclient==2.1.1  # For MySQL

# Authentication
dj-rest-auth==5.0.2
django-allauth==0.58.2

# API Documentation
drf-yasg==1.21.6

# Utilities
Pillow==10.0.1
python-dateutil==2.8.2
requests==2.31.0

# Development
pylint-django==2.5.3
pylint==3.0.2
black==23.9.1
isort==5.12.0

# Testing
pytest==7.4.2
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0
